Adds options hash to method call on resource
**Why**
The only way to configure OTP delivery options is with external
attributes which must persist on the model since sent_new_otp is called
outside of the request. This adds more flexibility in implementation of the
send_two_factor_authentication_code method.

**How**
Have the gem pass the options hash back to the send_two_factor_authentication_code
method on the resource. This permits at run-time options to reside with the send_new_otp
call. `