Allow auto-sending of OTP to be configurable
**Why**
The gem is built to hook in to Warden and automatically send the user the
OTP once authenticated. This is not always desired behavior.

**How**
Enable default behavior which sends OTP automatically in the after_authentication
Warden hook. Add configuration option for users of the gem to disable the automatic
transmission of OTP after authentication.